[NWS-1512][NWS-1826] Migrate ongoing contracts report generation
* Migrate ongoing contracts report

* Remove old generate report button

* Remove zero unit in styles

* Fix missing nb stepper provider

* Improve buttons color

* Change buttons width to half each

* Remove unused code

* Fix parameter in percentage functions calls

* Fix label for attribute
