+++ src/app/pages/contracts/contracts.component.html
@@ -3,12 +3,7 @@
     <nb-card>
       <nb-card-header *ngIf="pageWidth() < 576">
         <div class="row">
-          <div class="col-2" style="padding-left: 0; padding-right: 0">
-            <button nbButton fullWidth status="primary" style="height: 100%" (click)="openTeamDialog()">
-              <i class="icon-file-csv" style="font-size: 1.5rem"></i>
-            </button>
-          </div>
-          <div class="col-10">
+          <div class="col-12">
             <input
               nbInput
               [(ngModel)]="searchQuery"
@@ -36,7 +31,6 @@
           #smartTable
           [settings]="settings"
           [source]="source"
-          (create)="openTeamDialog()"
           (edit)="contractDialog($event, true)"
           (delete)="contractDialog($event, false)"
         ></nb-smart-table>
