+++ src/app/pages/contracts/contracts.component.scss
@@ -1,35 +1,6 @@
 @import 'app/@theme/styles/themes';
 
-@font-face {
-  font-family: 'icons';
-  src: url('../../../assets/fonts/icons.eot?6rw0ga');
-  src: url('../../../assets/fonts/icons.eot?6rw0ga#iefix') format('embedded-opentype'),
-    url('../../../assets/fonts/icons.ttf?6rw0ga') format('truetype'),
-    url('../../../assets/fonts/icons.woff?6rw0ga') format('woff'),
-    url('../../../assets/fonts/icons.svg?6rw0ga#icons') format('svg');
-  font-weight: normal;
-  font-style: normal;
-  font-display: block;
-}
-
 ::ng-deep {
-  .icon-file-csv {
-    font-family: 'icons' !important;
-    font-style: normal;
-    font-weight: normal;
-    font-variant: normal;
-    text-transform: none;
-    line-height: 1;
-
-    /* Better Font Rendering =========== */
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-  }
-
-  .icon-file-csv::before {
-    content: '\e900';
-  }
-
   nb-card {
     overflow-x: auto;
     overflow-y: auto;
