+++ src/app/pages/dashboard/report-menu/ongoing-contracts-report/ongoing-contracts-report.component.scss
@@ -1,0 +1,3 @@
+.stepper-button {
+  margin: 2rem 2rem 0 0;
+}
